Feature neural network parser

+++ ImplementationPhase/Model/NeuralNetworkModelModule/src/ConvolutionLayer.cpp
@@ -2,9 +2,9 @@
 
 ConvolutionLayer::ConvolutionLayer() : NetworkLayer(LayerType::CONVOLUTION), stride(0), padding(0) {}
 
-ConvolutionLayer::ConvolutionLayer(std::string name, int* inputDimensions, TENSOR(double) weights, int stride, int padding) : NetworkLayer(name, inputDimensions, LayerType::CONVOLUTION), weightsTensor(weights), stride(stride), padding(padding) {}
+ConvolutionLayer::ConvolutionLayer(std::string name, int* inputDimensions, TENSOR(float) weights, int stride, int padding) : NetworkLayer(name, inputDimensions, LayerType::CONVOLUTION), weightsTensor(weights), stride(stride), padding(padding) {}
 
-void ConvolutionLayer::setWeightsTensor(TENSOR(double) weights)
+void ConvolutionLayer::setWeightsTensor(TENSOR(float) weights)
 {
   weightsTensor = weights;
 }
@@ -19,7 +19,7 @@ void ConvolutionLayer::setPadding(int padding)
   this->padding = padding;
 }
 
-TENSOR(double) ConvolutionLayer::getWeightsTensor()
+TENSOR(float) ConvolutionLayer::getWeightsTensor()
 {
   return weightsTensor;
 }
